LambdaParser

Description

Copyrigth : (c) Nicolas Osborne, 2018 Licence : MIT

Provide the function pg2ast which turn a program written in Simply Typed Lambda Calculus into an Abstract Syntax Tree.

A program in Simply Typed Lambda Calculus is divided into two main parts:

An Assignations part containing assignations instruction of the form:

let id Exp

Where id is an identificator and Exp is a simply styped lambda expression.

A Program part containing the actual computation which may contain id previously defined in the Assignations part.

Documentation

data AST

Abstract Syntax Tree type declaration.

to be completed

Constructors

Empty

Leaf

Fields

len :: Int

length of the leaf, which is obviously 1

lambdatype :: [Char]

One of the defined type of the language

name :: [Char]

Either a value or an id

pg2ast :: [Char] -> AST

Turn a source file into the corresponding AST.

A source file must contain two main parts deliminated by the three following key words in that order: Assignations, Program, End. The assignation part (between the two first key words) is a series of assignation instruction (with the let key word) and the program part (between the two lasts key words) is the actual computation of which the source file is a description. The program part may contain the identificators declared in the assignation part.

input

The content of the source file.

output

the corresponding AST.

file2list :: [Char] -> [[Char]]

Turn a String into the corresponding list of tokens.

Split the string and drop the separators. Separators are whitespace and end of line.

extractAssignations :: [[Char]] -> [[Char]]

Return the sub-list corresponding to the list of tokens corresponding to the Assignation part of the source file.

extractPg :: [[Char]] -> [[Char]]

Return the sub-list corresponding to the list of tokens corresponding to the Program part of the source file

parseA :: [[Char]] -> [[[Char]]]

Parse the Assignation part of the source file.

Input

The list of tokens corresponding to the Assignation part of the source file.

Output

A list of assignations. The head of each item is an id and the tail is the expression

preparePg :: [[[Char]]] -> [[Char]] -> [[Char]]

Substitute the Id in the list of tokens corresponding to the program part of the source file by the corresponding list of tokens.

createAST :: [[Char]] -> AST

Create the AST corresponding to the given list of tokens.