Utils

Description

Documentation

oneSplit :: Char -> [Char] -> ([Char], [Char])

Split a string into two substrings, the prefix up to the first occurrence of the given Char, and the suffix beginning after this occurrence.

input

the separator

input

the string

output

the pair of the two substrings

stripPar :: [Char] -> [Char]

strip a string from its outer parenthesis if there is any