LambdaParser

Description

Copyrigth : Nicolas Osborne, 2018 Licence : MIT

Provide the function pg2ast which turn a program written in Simply Typed Lambda Calculus into an Abstract Syntax Tree.

Documentation

data Ast

Abstract Syntax Tree type declaration.

An Ast is either an empty tree (correspond to an empty program), or a Node with two children which are Abstract Syntax Trees with leaves. Nodes are for lambda application, so they have an arity of two. Leaves are for names.

Constructors

Empty

Node

Fields

len :: Int

length of the tree

operator :: [Char]

tag for the operator, either App or Abs

left :: Ast

either a binder or a lambda abstraction

rigth :: Ast

a lambda expression

Leaf

Fields

len :: Int

length of the tree

name :: [Char]

name: id:Type

pg2ast :: [Char] -> Ast

Turn a program into the corresponding Abstract Syntax Tree.

parsePg :: [Char] -> [[Char]]

Turn a program into the corresponding list of its tokens.

createAst :: [[Char]] -> Ast

Turn a list of the tokens of a program into the corresponding Abstact Syntax Tree.

readName :: [Char] -> [Char]

Extract the string corresponding of a name from the point of the program we have reached in the parsing.

readId :: [Char] -> [Char]

Check that the id is lower case.

readTypeJgt :: [Char] -> [Char]

Check there is a colon between id and type.

readType :: [Char] -> [Char]

Check the type is a known type.